move the types to their own files
