+++ DSharpPlus/Entities/User/DiscordActivity.cs
@@ -383,7 +383,7 @@ public sealed class DiscordRichPresence
             }
             else if (ulong.TryParse(lid, NumberStyles.Number, CultureInfo.InvariantCulture, out ulong ulid))
             {
-                this.LargeImage = new DiscordApplicationAsset { Id = lid, Application = this.Application, Type = ApplicationAssetType.LargeImage };
+                this.LargeImage = new DiscordApplicationAsset { Id = lid, Application = this.Application, Type = DiscordApplicationAssetType.LargeImage };
             }
         }
 
@@ -396,7 +396,7 @@ public sealed class DiscordRichPresence
             }
             else if (ulong.TryParse(sid, NumberStyles.Number, CultureInfo.InvariantCulture, out ulong usid))
             {
-                this.SmallImage = new DiscordApplicationAsset { Id = sid, Application = this.Application, Type = ApplicationAssetType.LargeImage };
+                this.SmallImage = new DiscordApplicationAsset { Id = sid, Application = this.Application, Type = DiscordApplicationAssetType.LargeImage };
             }
         }
     }
